bad things found
the duplication of cards is still at large. And I can't for the life of me figure out a way of making the knocking work well for the computer AND the player. My solutions are really only good for one of them. Gonna have to revamp. Maybe the scrapping of this version of the GameWindow class begins now?? (!!)